requirements.txt
blinker
cachelib
certifi
charset-normalizer
click
colorama
contourpy
cycler
Flask
Flask-Caching
Flask-Cors
Flask-JWT-Extended
Flask-SSE
fonttools
geographiclib
geopy
gunicorn
idna
itsdangerous
Jinja2
kiwisolver
MarkupSafe
matplotlib
numpy
packaging
pandas
patsy
Pillow
pip
protobuf
psycopg2-binary
PyJWT
pyparsing
python-dateutil
python-dotenv
pytz
redis
requests
scipy
setuptools
six
statsmodels
tzdata
ujson
urllib3
Werkzeug